Downloader: Make DownloadFile(), RedownloadSlices() and CheckCurrentFile() asynchronous
This prepares Downloader.cs code in full to make the slices parallelizable easier.

The delays/sleeps arent needed since we will have some threads overhead and we're waiting in the code anyway.

The code works faster now albeit it seems to download when it says "finished" instead of "started", but this could be how it should be, considering the awaits.

Slightly faster this way, but still not parallel yet.